Fixing endian checking code for 64bit system, fixed first by Hantao and (ver unfortunately) again by Chris Kelly.
git-svn-id: file:///home/chulwoo/CPS/repo/CVS2SVN/cps/trunk@4050 fb734a29-28db-4bc5-88a3-3cb5ee595545

+++ cps_pp/src/util/qcdio/comsrc/intconv.C
@@ -103,7 +103,7 @@ enum INT_FORMAT  IntConv::testHostFormat() { // test the type of CPS::Float
   const char * fname = "testHostFormat()";
   // 1. endian
   char end_check[4] = {1,0,0,0};
-  unsigned long *lp = (unsigned long *)end_check;
+  uint32_t *lp = (uint32_t *)end_check;
   int host_big;
 
   if ( *lp == 0x1 ) {
